requirements.txt
python-telegram-bot==20.0
motor==3.1.1
python-dotenv==0.19.2
aiohttp==3.8.1